refactor(catalog): extract foundation-pure leaves to proximadb-catalog (Slice 1)
First increment of the catalog/metrics root-crate extraction. Moves 9
foundation-pure files from src/catalog into the proximadb-catalog crate:
corpus_version_fs_store, internal/{mod,enforcement,information_schema,registry},
partition_pruning, recall_probe, syscat_cache, syscat_warm.

The catalog crate already holds the backends; root src/catalog is the
runtime/wiring layer. These 9 files depend only on foundation + intra-clean +
the Catalog trait, so they move cleanly (the observability Slice-1 pattern).

Crate changes: +deps proximadb-cache, proximadb-filter-expression,
proximadb-proto, chrono, prost; +pub mod proto re-export (mirrors the
observability engine crate). Root src/catalog/mod.rs re-exports the moved
modules (pub use proximadb_catalog::X;) so all crate::catalog::* import paths
are unchanged; keeps CatalogManager + the deferred modules.

Deferred (stay root, later slices): segment_registry (needs
storage_common::pax_block::SegmentMeta -- catalog->storage-common would cycle on
the existing downward edge); budget_guard + tier_transition (use tenant_tier);
federation + iceberg_rest_service (use CatalogManager); tenant_tier /
tier_recommendation / index_location_resolver (coupled to query / observability
/ index).

Plan doc: ROOT_CRATE_DECOMPOSITION_CATALOG_METRICS_2026_07_12.adoc (catalog low
coupling ~5 up-edges first; metrics has 10 storage up-edges and needs a
port-inversion -- the larger, later slice).

Storage up-edge ratchet unchanged at 197. crate + root clippy --lib --bins
-D warnings clean; layering clean; 230 catalog lib tests pass.
