+++ crates/control/proximadb-catalog/Cargo.toml
@@ -29,6 +29,10 @@ unity-catalog         = ["dep:reqwest", "dep:urlencoding"]
 anyhow.workspace = true
 arrow-schema = { workspace = true, default-features = false }
 async-trait.workspace = true
+# Extracted runtime/wiring modules (Slice 1) use chrono (timestamps) + prost
+# (proto message construction in syscat_cache/warm).
+chrono.workspace = true
+prost.workspace = true
 aws-config = { version = "1.0", optional = true }
 aws-sdk-glue = { version = "1.0", optional = true }
 dashmap.workspace = true
@@ -40,6 +44,14 @@ proximadb-data-model = { workspace = true }
 proximadb-distance-types = { workspace = true }
 proximadb-quantization-types = { workspace = true }
 proximadb-records = { workspace = true }
+# Foundation deps pulled in by the runtime/wiring modules extracted from the
+# root `src/catalog` (Slice 1): proto messages (Collection), filter expressions
+# (partition pruning), and the cache façade. (segment_registry stays root — it
+# needs storage_common::pax_block::SegmentMeta, and storage-common→catalog is a
+# downward edge, so catalog→storage-common would cycle.)
+proximadb-cache = { workspace = true }
+proximadb-filter-expression = { workspace = true }
+proximadb-proto = { workspace = true }
 # FileSystem trait only (no concrete backends) — lets NativeCatalog persist
 # through an injected object-store/local backend (dependency inversion; the
 # concrete FilesystemFactory lives in the root crate and injects the impl).
