+++ docs/12-design/ROOT_CRATE_DECOMPOSITION_CATALOG_METRICS_2026_07_12.adoc
@@ -1,0 +1,96 @@
+= Root-Crate Decomposition — Catalog & Metrics Extraction (2026-07-12)
+:toc: macro
+:icons: font
+
+[abstract]
+Companion to `ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.adoc` and the CI-build
+optimization plan (#876). Scopes extraction of the two remaining non-storage
+root subsystems flagged by the observability handoff — `src/catalog` and
+`src/metrics` — behind the existing `proximadb-catalog` crate (catalog) and a
+future metrics crate. Uses the same measured-coupling / port-inversion approach
+that landed observability Slices 1–3 (#878/#879/#887).
+
+toc::[]
+
+== Why these two
+
+The observability program (~40K syms across 3 slices) is merged. The handoff
+named "catalog/metrics" as the remaining clean-slice work. Measured on
+`origin/develop` (2026-07-12):
+
+[cols="2,1,1,1",options="header"]
+|===
+| Subsystem | Root footprint | Up-edges (→ other root subsystems) | Difficulty
+| `src/catalog` | 19 runtime files (~21K syms) | **~5** (`core`×2, `storage`×1, `observability`×1, `index`×1) | **Low**
+| `src/metrics` | ~40 files (~25K syms) | **10 into `storage`** + `graph`×1 | **High** (needs a storage port-inversion)
+|===
+
+Catalog is the clear first target: low coupling, and the `proximadb-catalog`
+crate already exists with the backends (`delta`/`glue`/`polaris`/`unity`/
+`iceberg`/`corpus_version`/`native`/`relational`/…). Root `src/catalog` is the
+runtime/wiring layer that belongs there.
+
+NOTE: the graph-prioritized CI doc (#876) ranks **storage** extraction as the
+higher-ROI lever (287K syms, 34% of the monolith) and frames catalog as
+"already a crate → stabilize API." Catalog/metrics is still real, valuable
+root-shrink (~46K syms combined); it is the *clean-slice* track, complementary
+to (not instead of) the storage track.
+
+== Catalog — per-file coupling (measured)
+
+Non-foundation up-edges per file (`grep use crate::(storage|query|services|
+index|network|cluster|cdc|graph|observability|compute)::`):
+
+- *Coupled core (defer)* — `index_location_resolver` (index+storage),
+  `tenant_tier` (query), `tier_recommendation` (observability).
+- *Depend on root runtime (defer)* — `budget_guard` + `tier_transition` (both
+  use `tenant_tier::*` types); `federation/federated_catalog` +
+  `iceberg_rest_service` (both use `CatalogManager`, defined in root `mod.rs`).
+- *`CatalogManager` itself* (root `mod.rs:216`) is the central runtime — it
+  stays root until a later "facade" slice (the observability-Slice-3
+  equivalent), possibly behind a port.
+
+== Catalog Slice 1 (this PR) — foundation-pure leaves
+
+Move the 9 foundation-pure files into `proximadb-catalog`:
+
+- `corpus_version_fs_store.rs` (deps `corpus_version`, already in crate)
+- `internal/{mod,enforcement,information_schema,registry}.rs` (self-contained)
+- `partition_pruning.rs` (`core::search::{FilterExpression,ComparisonOperator}`
+  → `proximadb-filter-expression`, a foundation re-export)
+- `recall_probe.rs` (no deps)
+- `syscat_cache.rs`, `syscat_warm.rs` (need proto `Collection` → add
+  `proximadb-proto` dep + proto re-export module)
+
+Crate changes: add deps `proximadb-cache`, `proximadb-filter-expression`,
+`proximadb-proto`, `chrono`, `prost` + a `pub mod proto` re-export (mirrors the
+observability crate). Root `src/catalog/mod.rs` keeps `CatalogManager` + the
+deferred modules and re-exports the moved items (`pub use proximadb_catalog::X;`)
+so all `crate::catalog::*` import paths are unchanged.
+
+NOTE: `segment_registry` was *deferred* — it needs
+`storage_common::pax_block::SegmentMeta`, and `storage-common → catalog` is
+already a downward edge, so `catalog → storage-common` would form a cycle.
+It stays root until `SegmentMeta` is hoisted to a shared foundation crate.
+
+== Sequencing
+
+. **Catalog Slice 1** — 10 foundation-pure leaves (this PR).
+. **Catalog later slices** — extract `CatalogManager` + the coupled core behind
+  a port (invert the index/storage/query/observability up-edges); then the
+  deferred files follow.
+. **Metrics** — new `proximadb-metrics` (or fold into `proximadb-telemetry`);
+  the 10 `storage` up-edges need a `MetricsStoragePort` port-inversion (the
+  observability-Slice-3 pattern). Largest single piece.
+
+== Reproducibility
+
+[source,bash]
+----
+# per-file non-foundation up-edges
+for f in $(find src/catalog -name '*.rs'); do
+  e=$(grep -hE "use crate::(storage|query|services|index|network|cluster|cdc|graph|observability|compute)::" "$f" \
+      | sed -E 's/.*(crate::[a-z_]+)::.*/\1/' | sort -u | tr '\n' ',' | sed 's/,$//')
+  [ -n "$e" ] && echo "COUPLED [$e]: $f" || echo "clean: $f"
+done
+----
